refactor: enhance file handling and selection logic in table component
