+++ apps/data-download-portal/src/pages/dataset/dataset.tsx
@@ -1,10 +1,9 @@
 import { Fragment, useEffect, useState } from 'react'
 import { useParams } from '@tanstack/react-router'
-import { isNumber } from 'lodash'
 import { DateTime } from 'luxon'
 
 import { GFWAPI } from '@globalfishingwatch/api-client'
-import type { Dataset, DatasetFile } from '@globalfishingwatch/api-types'
+import type { Dataset } from '@globalfishingwatch/api-types'
 import { IconButton } from '@globalfishingwatch/ui-components/icon-button'
 
 import ApiBanner from '../../components/api-banner/api-banner'
@@ -12,69 +11,10 @@ import Loader from '../../components/loader/loader'
 import EnhancedMarkdown from '../../components/markdown/markdown'
 import Table, { type TableData } from '../../components/table/table'
 import TopBar from '../../components/topBar/topBar'
-import { MAX_DOWNLOAD_FILES_LIMIT } from '../../config.js'
 import { getUTCString } from '../../utils/dates.js'
+import { buildFileTree, formatBytes } from '../../utils/folderConfig'
 
 import styles from './dataset.module.scss'
-
-function formatBytes(bytes: number, decimals = 2) {
-  if (bytes === 0) return '0 Bytes'
-
-  const k = 1024
-  const dm = decimals < 0 ? 0 : decimals
-  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB']
-
-  const i = Math.floor(Math.log(bytes) / Math.log(k))
-
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i]
-}
-
-const insertIntoTree = (
-  tree: TableData[],
-  parts: string[],
-  fullPath: string,
-  size: number,
-  lastUpdate: string
-) => {
-  let currentLevel = tree
-
-  for (let i = 0; i < parts.length; i++) {
-    const part = parts[i]
-    const isFolder = i < parts.length - 1 || fullPath.endsWith('/')
-
-    let existing = currentLevel.find((node) => node.name === part)
-
-    if (!existing) {
-      const node: TableData = {
-        name: part,
-        path: parts.slice(0, i + 1).join('/'),
-        lastUpdate: lastUpdate ? DateTime.fromISO(lastUpdate).toFormat('M/dd/yyyy') : '---',
-        size: size || '-',
-        ...(isFolder ? { subRows: [] } : {}),
-      }
-
-      currentLevel.push(node)
-      existing = node
-    }
-
-    if (existing.subRows) {
-      currentLevel = existing.subRows!
-    }
-  }
-}
-
-const buildFileTree = (files: DatasetFile[]): TableData[] => {
-  const tree: TableData[] = []
-
-  files.forEach((file) => {
-    const cleanPath = file.name.endsWith('/') ? file.name.slice(0, -1) : file.name
-    const parts = cleanPath.split('/')
-
-    insertIntoTree(tree, parts, file.name, Number(file.size), file.lastUpdate)
-  })
-
-  return tree
-}
 
 const columns = [
   {
@@ -84,33 +24,35 @@ const columns = [
     Cell: ({ row }: { row: any }) => (
       <span
         style={{
-          paddingLeft: `${row.depth * 1.3}rem`,
+          paddingLeft: `${row.depth * 1.5}rem`,
           display: 'flex',
           gap: '0.5rem',
           alignItems: 'center',
         }}
       >
-        {`${row.values.name}`}
-        {row.canExpand &&
-          (row.isExpanded ? (
-            <IconButton icon="arrow-top" size="small" type="invert" />
-          ) : (
-            <IconButton icon="arrow-down" size="small" type="invert" />
-          ))}
+        {row.values.name}
+        {row.canExpand && (
+          <IconButton
+            icon={row.isExpanded ? 'arrow-top' : 'arrow-down'}
+            size="small"
+            type="invert"
+          />
+        )}
       </span>
     ),
   },
   {
     id: 'size',
-    width: 60,
+    width: 55,
     Header: 'Size',
-    accessor: (row: any) => (isNumber(row.size) ? formatBytes(row.size) : '--'),
+    accessor: (row: any) => (Number(row.size) ? formatBytes(row.size) : '--'),
   },
   {
     id: 'date',
     width: 80,
     Header: 'Date',
-    accessor: (row: any) => row.lastUpdate,
+    accessor: (row: any) =>
+      row.lastUpdate ? DateTime.fromISO(row.lastUpdate).toFormat('M/dd/yyyy') : '---',
   },
 ]
 
@@ -161,16 +103,6 @@ function DatasetPage() {
               </div>
             </div>
             <div className={styles.files}>
-              {/* UNCOMMENT WHEN ALL FILES DOWNLOAD READY */}
-              {/* <label>Entire dataset</label>
-              <a
-                className="btn"
-                href={`${API_GATEWAY}/download/datasets/${datasetId}/download-single/all`}
-              >
-                Download dataset
-              </a>
-              <br /> */}
-
               {dataset && dataset.files && (
                 <Table columns={columns} data={dataset.files as TableData[]} />
               )}
